Separate PreferenceMatrix into matrix and feedback
While working on benchmarking, I noticed that PreferenceMatrix
construction takes up a significant amount of time during the algorithm
execution (python3 -m cProfile -s tottime duelpy/benchmarks/benchmarks.py).

That is because some PreferenceMatrices in PreferenceEstimate are
constructed without passing a RandomState. In that case, a new
RandomState is initialized which is slow. Those PreferenceMatrices don't
really need the random state at all: They are just used for utility
functions such as `get_condorcet_winner`, not for duels.

One solution would be to pass a random state to the preference estimate
and pass that on. But it seems like feedback.PreferenceMatrix is doing
too much. It implements a preference matrix with utility functions,
dueling and history / regret computation. This commit splits the
PreferenceMatrix out of the feedback/history. That way we can return the
more lightweight PreferenceMatrix class which does not require a random
state. This cuts down the benchmarking time for 1e5 samples and 100 arms
with savage from 70 seconds to 45 seconds on my machine.