Ignore import similarities
Pulled out of !45 since the issue was popping up on other MRs. The
similar import check of pylint goes a bit too far. Similar imports can
of course hint at similar code, but that is by design in our case since
we implement many algorithms that fundamentally do the same / a similar
thing.