Optimize preference matrix sample
Going through the matrix in a for loop is inefficient. Taking individual
samples instead of sampling all elements at once is inefficient as well.

As a result, this function took up a significant fraction of the
algorithm runtime. By using matrix operations and batch sampling
instead, the amount of time is now almost negligible.