Optimize confidence radius caching
A first stab at slow runtimes due caused by cache invalidation when
changing the confidence radius, as described in [1]. On a high level,
the changes are:

- Instead of partial cache invalidation, the cache is now always
  completely invalidated. The idea is to reduce book-keeping, but this
  might be re-evaluated in the future.
- When the cache is recomputed, the samples are de-duplicated. Since
  there are many pairs of arms, its likely that a lot of them share the
  same number of samples (though this depends on the exact
  implementation of the algorithm of course).

Its probably possible to do better, but this gets us pretty close to the
"always cached" case already. A further optimization would be to replace
the computation of the confidence radii for the unique num_sample values
with a vectorized numpy operation. That should be relatively
straightforward now, if it still turns out to be a bottleneck. An
alternative would be to look into partial invalidation again. This could
be combined with a "dirty" flag for the cache, to balance book-keeping
(only a single flag must be checked most of the time) with caching.

I measured the runtime of our current algorithms with the
"savage_experiment" environment, 20 arms and a time horizon of 10000:

                               before      after  delta
algorithm
BeatTheMeanBandit            8.759201   8.073279    0.7
BeatTheMeanBanditPAC        10.726643  10.688775    0.0
CopelandConfidenceBound     59.853995  19.297551   40.6
DoubleThompsonSampling      56.804992  11.869105   44.9
InterleavedFiltering         0.403315   0.394197    0.0
KnockoutTournament           0.281153   0.272874    0.0
MallowsMPI                   0.420037   0.418397    0.0
MallowsMPR                   0.527453   0.594792   -0.1
RelativeConfidenceSampling   6.043051   4.848157    1.2
RelativeUCB                 56.316095   9.262959   47.1
Savage                       0.801803   0.773769    0.0
SequentialElimination        0.344400   0.328521    0.0
WinnerStaysStrongRegret      0.107561   0.105273    0.0
WinnerStaysWeakRegret        1.849139   1.810212    0.0

[1] https://gitlab.com/duelpy/duelpy/-/issues/29#note_456039647