Avoid repeated Copeland score computations
This inlines the average Copeland score computation into DTS+ and pulls
the normalized Copeland score computation out of the loop. The caching
could also happen within the PreferenceMatrix, but since DTS+ is
currently the only user of that function it seems easiest and best to
inline it for now. We can reconsider this once another algorithm needs
that functionality.

As a result, the performance of DTS+ has already improved by quite a
bit:

$ python3 duelpy/experiments/savage_experiment.py --arms 20 --time-horizon 10000 --runs 2 --profile -a DoubleThompsonSamplingPlus

yields

                                     wall_clock
DoubleThompsonSamplingPlus (before)  78.527847
DoubleThompsonSamplingPlus (after)   30.931789