Test the PreferenceMatrix with a changing radius
The previous test did not catch an issue with the new caching strategy.
The confidence radius is not invalidated when a new sample is entered,
which didn't fail the test since the confidence radius was constant
anyway. The new test should catch that kind of issue and is also a
better illustration of the functionality.