Invalidate the radius cache on a new sample
This fixes a regression caused by !92. The old tests didn't catch this,
but the new tests (see 005427db13b1ec7c9b8eadce69005872f09bc8e9) should
make sure that such a regression does not happen again.

We have to re-compute the whole cache, even if only a single value has
changed since it was last computed. That is a deliberate choice to avoid
book-keeping overhead.  We may have to re-evaluate that in the future if
it turns out that re-computation is more expensive than the book-keeping
would be.