Use the metrics module in the experiment
This is a step towards disentangling the experiment code. The metrics
are now defined in their own module. That makes is easy to add
additional metrics or choose a subset of the existing metrics for an
experiment. The goal is to eventually have a set of components (such as
experiment runner, metrics, plotting and data collection) that can be
put together in different ways to create different experimental setups.

It also improves the plots a bit, since samples are now taken after
every duel instead of every "logical step". Most metrics were already
recorded per-duel by the MatrixFeedback class, but the wall clock time
wasn't.