Reuse the same random state in winner stays
Instantiating the two different random states with the same seed can
lead to unintended behavior. It will generate the same sequence of
random numbers for each instance independently. That doesn't matter much
in this example, since MatrixFeedback and WinnerStays{,WeakRegret} use
the random state for different things. Still, I think it is a bit
cleaner to reuse the same random state.