Make the winner stays example more challenging
A regret of 0 in the doctest looks a bit suspicious. Ideally the regret
values in the doctest can help us catch regressions, but that is less
likely to work if the example is too simple. (It would be good if we had
a suite of benchmarks in addition to the doctests).

A weak regret of 0 is reasonable in this case: All WinnerStays has to do
is guess the optimal arm (1/3) for the first duel and then stick to it.

This change makes the "stick to it" step a bit harder by making the
Condorcet winner harder to distinguish from the other arms.