Use the superclass implementation of step
This function is defined in the PacAlgorithm superclass. Most
implementations were equivalent to the default implementation, but some
small adjustments were necessary.

The logic has changed a bit for Mallows and Plackett-Luce, since the
early return is not included in the default implementation. I think that
is reasonable, since step should not be called in the first place if
`is_finished` returns true.

The change in the way OptMax uses sequential elimination is necessary
since the AlgorithmFinishedException is now caught during exploration.
The way this works is not ideal and will likely be changed in the
future.