Implement exploitation for result interfaces
PAC algorithms that produce the same result usually use the same kind of
exploitation.

The order of inheritance matters if you want to use these exploit
functions. Those classes that are later in the list take precedence.
It would be better to avoid this slightly confusing behavior of multiple
inheritance. An alternative would be to have all the "Producer"
superclasses inherit from PacAlgorithm. I have not done it that way
since currently we have some algorithm implementations that inherit from
a Producer superclass but are not PAC algorithms.