Explicitly set common color and style palette
These were implicitly the same already, but setting them explicitly is
more reliable. The old way could break if seaborn changes its internals
or if we add an additional plot that e.g. uses only a subset of the
algorithms.

This also restricts the dash styles to two different kinds. This is
simpler to implement (`sns._core.unique_dashes` exists but is a seaborn
internal) and also makes the plots look less "noisy" to me. The styles
are redundant anyway, so it is not necessary to have a distinct style
per line.