Format the sphinx configuration with black
This is not enforced by pre-commit due to the global exclude in
`.pre_commit_commit_config.yaml`. Some other linters are still failing
on the `sphinx` subdirectory. Arguably most of them are not so important
for the sphinx configuration, but disabling `doc8` for that folder
pretty much disables it entirely. It is out of scope for this commit to
change that though.