Use a decorator in the decision experiment
Using `step` to print the preference estimate after each duel only works
if `step` actually only performs one duel. That is the case for Savage,
but not for many other algorithm implementations. The example is quite
a bit longer now, but most of it is documentation. It gives a better
insight into the library but the downside is that it might be a bit
daunting as a first introduction.