Use a budgeted feedback mechanism by default
This should simplify early termination in many algorithms. The feedback
mechanism is wrapped by default. It will raise an exception once the
budget is exceeded (if one is given). The exception is caught in the
`run` functions. In most cases the algorithm itself will not need to
implement any special behavior. The wrapper can be stacked and only the
exception for this particular budgeted feedback instance is caught, so
this should work nicely with nested algorithms.

This effectively deprecates the option of calling `step` manually. It
can still be done, but one needs to be aware that it may raise an
exception. We could introduce another layer of indirection instead
(`step` calls `_step` or `_logical_step`), but I think it is not worth
it. The `step` function may execute an arbitrary number of duels and
therefore is not really useful as a "unit of computation" after which
one might want to inspect the algorithm's internal state. You should use
a feedback mechanism decorator instead, which allows you to inspect it
after every duel. See `duelpy/examples/decision.py` for an example.

The wrapped feedback might lead to unexpected behavior. I plan to rename
the feedback mechanism in a follow-up commit to clarify that it is
wrapped.