Refactor OptMax and SequentialElimination
This uses the "PAC mode" of Sequential elimination and runs it to
completion (if it is not terminated by the `BudgetedFeedbackMechanism`
decorator) instead of using the "regret minimizing mode" and then only
calling `explore`. I think that is easier to understand. It no longer
relies on implementation details of `SequentialElimination`. We do not
need to manually raise `AlgorithmFinishedExceptions` any more. The
wrapped feedback mechanism already raises an exception if a duel would
exceed a given time horizon.