Do not pass a time horizon to top-1 selection
The previous implementation would use up the entire duel budget on the
first top-1 selection run. If a time horizon is given, top-1 selection
would use exploitation to make sure it is reached. That is not the
desired behavior. Top-1 selection should execute until it has determined
a result. The wrapped feedback mechanism ensures that the top-1
selection cannot exceed the time horizon of top-k sorting.