Do not terminate KLDBP exploration early
The check was likely intended to ensure that the time horizon is not
exceeded. It is no longer necessary for that purpose since the feedback
mechanism is wrapped with a `BudgetedFeedbackMechanism` decorator.

If a time horizon was not given, `is_finished` would be equivalent to
`exploration_finished`. That means that the reward generation could be
terminated before all rewards have been generated. That is not what the
algorithm should do.